+++ client/src/components/Upload/index.js
@@ -6,3 +6,5 @@ export { default as UploadModal } from "./UploadModal";
 export { initializeUploadDefaults } from "./config";
 
 export { mount as mountUploadModal } from "./mount";
+
+export { default as UploadButton } from "./UploadButton";
